revival

+++ final_project/src/main/java/pl/coderslab/final_project/entity/user/Wallet.java
@@ -1,7 +1,5 @@
 package pl.coderslab.final_project.entity.user;
 
-import java.math.BigDecimal;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -21,7 +19,7 @@ public class Wallet {
 	@OneToOne
 	private User owner;
 	@Digits(fraction=2, integer = 12)
-	private BigDecimal accState;
+	private double accState;
 	
 	
 }
